FileMatcherTest: Move several tests to RootLicenseMatcherTest
In order to prepare for the removal of the now more needed
FileMatcher.LICENSE_FILE_MATCHER.
